Modify package structure
