+++ app/src/main/java/com/trx/consumer/common/CommonLiveEvent.kt
@@ -5,23 +5,27 @@ import androidx.annotation.Nullable
 import androidx.lifecycle.LifecycleOwner
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.Observer
-import java.util.concurrent.atomic.AtomicBoolean
 import timber.log.Timber
+import java.util.concurrent.atomic.AtomicBoolean
 
-class SingleLiveEvent<T> : MutableLiveData<T>() {
+class CommonLiveEvent<T> : MutableLiveData<T>() {
 
     private val pending = AtomicBoolean(false)
 
     @MainThread
     override fun observe(owner: LifecycleOwner, observer: Observer<in T>) {
         if (hasActiveObservers()) {
-            Timber.tag(SingleLiveEvent::class.java.simpleName).w("Multiple observers registered but only one will be notified of changes.")
+            Timber.tag(CommonLiveEvent::class.java.simpleName)
+                .w("Multiple observers registered but only one will be notified of changes.")
         }
-        super.observe(owner, Observer {
-            if (pending.compareAndSet(true, false)) {
-                observer.onChanged(it)
+        super.observe(
+            owner,
+            {
+                if (pending.compareAndSet(true, false)) {
+                    observer.onChanged(it)
+                }
             }
-        })
+        )
     }
 
     @MainThread
