+++ app/src/main/java/com/trx/consumer/managers/CacheManager.kt
@@ -1,0 +1,10 @@
+package com.trx.consumer.managers
+
+import javax.inject.Inject
+
+class CacheManager @Inject constructor() {
+
+    fun isUserLoggedIn(): Boolean {
+        return false
+    }
+}
