+++ app/src/main/java/com/trx/consumer/screens/splash/SplashFragment.kt
@@ -1,38 +1,35 @@
-package com.trx.consumer.module.splash
+package com.trx.consumer.screens.splash
 
 import android.os.Bundle
 import android.view.View
 import androidx.fragment.app.activityViewModels
 import androidx.lifecycle.Observer
 import com.trx.consumer.R
-import com.trx.consumer.core.BaseFragment
-import com.trx.consumer.core.MainActivity
+import com.trx.consumer.base.BaseFragment
 import kotlinx.android.synthetic.main.fragment_splash.*
 
 class SplashFragment : BaseFragment(R.layout.fragment_splash) {
 
+    //region Objects
     private val viewModel: SplashViewModel by activityViewModels()
+    //endregion
 
+    //region Setup
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-        initViewModel()
-        initClickListeners()
-        initViewModelObservers()
-    }
-
-    private fun initViewModel() {
         viewModel.onLoadView()
+        bind()
     }
 
-    private fun initViewModelObservers() {
+    private fun bind() {
         viewModel.eventLoadView.observe(viewLifecycleOwner, handleLoadView)
         viewModel.eventTapStart.observe(viewLifecycleOwner, handleTapStart)
-    }
 
-    private fun initClickListeners() {
         btnStart.setOnClickListener { viewModel.onTapStart() }
     }
+    //endregion
 
+    //region Handlers
     private val handleLoadView = Observer<Void> {
         tvTitle.text = getString(R.string.splash_title)
         tvSubtitle.text = getString(R.string.splash_subtitle)
@@ -40,4 +37,5 @@ class SplashFragment : BaseFragment(R.layout.fragment_splash) {
     }
 
     private val handleTapStart = Observer<Void> { }
+    //endregion
 }
