--- app/src/main/java/com/trx/consumer/services/user/UserDataSource.kt
@@ -1,6 +1,0 @@
-package com.trx.consumer.services.user
-
-interface UserDataSource {
-
-    fun isUserLoggedIn(): Boolean
-}
