--- app/src/main/java/com/trx/consumer/services/user/UserRepository.kt
@@ -1,10 +1,0 @@
-package com.trx.consumer.services.user
-
-import javax.inject.Inject
-
-class UserRepository @Inject constructor() : UserDataSource {
-
-    override fun isUserLoggedIn(): Boolean {
-        return false
-    }
-}
