+++ app/src/test/java/com/trx/consumer/SplashViewModelTests.kt
@@ -2,7 +2,7 @@ package com.trx.consumer
 
 import androidx.arch.core.executor.testing.InstantTaskExecutorRule
 import androidx.lifecycle.Observer
-import com.trx.consumer.module.splash.SplashViewModel
+import com.trx.consumer.screens.splash.SplashViewModel
 import org.junit.After
 import org.junit.Before
 import org.junit.Rule
